Created a Working Wizard Class
Wizards have been given their own class and parameters.  They are the only unit spawning in the WHITE faction, as their image on the form is W/W

I basically only allowed melee and ranged to be permitted in faction index 0 and 1, whilst anything in 2 is forced to become a wizard.